fix OracleValue data conversion
* store all Oracle Price data as integer data

+++ runtime/development/src/lib.rs
@@ -499,7 +499,7 @@ impl kylin_oracle::Config for Runtime {
 
 		type CombineData = DefaultCombineData<Self, ConstU32<3>, ConstU128<600>>;
 	type OracleKey = Vec<u8>;
-	type OracleValue = f64;
+	type OracleValue = i64;
 	type Members = OracleProvider;
 	type MaxHasDispatchedSize = ConstU32<100>;
 }
